date formatter fall back to PHP impl. to solve Y2K38 issue
formatting date and time values for years >=2038 or <=1901 is now possible also with intl extension is
installed. 32bit systems will fall back to the PHP implementation because intl uses a 32bit UNIX timestamp internally.
On a 64bit system the intl formatter is used in all cases if installed.

fixes #5577

Conflicts:
	framework/CHANGELOG.md
	framework/i18n/Formatter.php
	tests/unit/framework/i18n/FormatterDateTest.php